poultryguard-ai/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
celery
redis
tensorflow
pillow
librosa
openai-whisper
openai
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
faiss-cpu
scikit-learn
scipy
reportlab
twilio
prometheus-fastapi-instrumentator
alembic
opencv-python
numpy
torch
tqdm
google-generativeai
langchain-google-genai
